@@ -11,8 +11,9 @@
 		"url": "https://sindresorhus.com"
 	},
 	"engines": {
-		"node": ">=8"
+		"node": ">=12.20"
 	},
+	"type": "module",
 	"scripts": {
 		"test": "xo && ava && tsd"
 	},
@@ -47,8 +48,9 @@
 		"trim-repeated": "^1.0.0"
 	},
 	"devDependencies": {
-		"ava": "^1.4.1",
-		"tsd": "^0.7.1",
-		"xo": "^0.24.0"
+		"ava": "^3.15.0",
+		"dirname-filename-esm": "^1.1.1",
+		"tsd": "^0.17.0",
+		"xo": "^0.40.3"
 	}
 }
