@@ -1,23 +1,21 @@
-declare namespace filenamify {
-	interface Options {
-		/**
-		String to use as replacement for reserved filename characters.
+export interface Options {
+	/**
+	String to use as replacement for reserved filename characters.
 
-		Cannot contain: `<` `>` `:` `"` `/` `\` `|` `?` `*`
+	Cannot contain: `<` `>` `:` `"` `/` `\` `|` `?` `*`
 
-		@default '!'
-		*/
-		readonly replacement?: string;
+	@default '!'
+	*/
+	readonly replacement?: string;
 
-		/**
-		Truncate the filename to the given length.
+	/**
+	Truncate the filename to the given length.
 
-		Systems generally allow up to 255 characters, but we default to 100 for usability reasons.
+	Systems generally allow up to 255 characters, but we default to 100 for usability reasons.
 
-		@default 100
-		*/
-		readonly maxLength?: number;
-	}
+	@default 100
+	*/
+	readonly maxLength?: number;
 }
 
 /**
@@ -25,7 +23,7 @@ Convert a string to a valid filename.
 
 @example
 ```
-import filenamify = require('filenamify');
+import filenamify from 'filenamify';
 
 filenamify('<foo/bar>');
 //=> 'foo!bar'
@@ -34,6 +32,4 @@ filenamify('foo:"bar"', {replacement: '🐴'});
 //=> 'foo🐴bar'
 ```
 */
-declare const filenamify: (string: string, options?: filenamify.Options) => string;
-
-export = filenamify;
+export default function filenamify(string: string, options?: Options): string;
