@@ -1,6 +1,6 @@
 import {expectType} from 'tsd';
-import filenamify = require('.');
+import filenamify, {filenamifyPath} from './index.js';
 
 expectType<string>(filenamify('<foo/bar>'));
 expectType<string>(filenamify('foo:"bar"', {replacement: '🐴'}));
-expectType<string>(filenamify.path('/some/!path'));
+expectType<string>(filenamifyPath('/some/!path'));
